requirements.txt
greenlet==1.1.0
mysqlclient==2.0.3
psycopg2-binary==2.8.6
pymongo==3.11.4
python-dotenv==0.17.1
six==1.16.0
SQLAlchemy==1.4.15
SQLAlchemy-Utils==0.37.4
